+++ extern/agidl/include/agidl_imgp_halftone.h
@@ -9,16 +9,16 @@
 *   Library: libagidl
 *   File: agidl_imgp_halftone.h
 *   Date: 1/27/2024
-*   Version: 0.2b
-*   Updated: 1/27/2024
+*   Version: 0.4b
+*   Updated: 6/9/2024
 *   Author: Ryandracus Chapman
 *
 ********************************************/
 
-#include "agidl_types.h"
-#include "agidl_cc_types.h"
-#include "agidl_img_types.h"
+#include <agidl_types.h>
+#include <agidl_cc_types.h>
+#include <agidl_img_types.h>
 
-void AGIDL_HalftoneImgData(void* data, u16 width, u16 height, AGIDL_CLR_FMT fmt, u8 threshold);
+void AGIDL_HalftoneImgData(void* data, u32 width, u32 height, AGIDL_CLR_FMT fmt, u8 threshold);
 
 #endif
